requirements.txt
requests
beautifulsoup4
tldextract
pillow